Restructure the folder structure
