+++ src/redux/crypto/actionTypes.js
