+++ src/redux/crypto/cryptoReducer.js
@@ -1,0 +1,20 @@
+import { LOAD_SUCCESS } from './actionTypes';
+
+const initialState = {
+  loading: true,
+  crypto: [],
+};
+
+const cryptoReducer = (state = initialState, action) => {
+  switch (action.type) {
+    case LOAD_SUCCESS:
+      return {
+        loading: false,
+        crypto: action.payload,
+      };
+    default:
+      return state;
+  }
+};
+
+export default cryptoReducer;
